content/loader/requirements.txt
sqlmodel>=0.0.14
psycopg[binary]>=3.1
python-dotenv>=1.0
